COM-1695: big refacto
