+++ src/screens/ImagePickerManager/styles.ts
